chore: change to non-free license
Unlike our bazel-lib library, which is needed pretty much everywhere in the Bazel ecosystem, and unlike rules_js, which most Bazel users require just to get started, rules_aws is a 'last mile' integration for larger organizations.
Therefore we aren't obligated to provide it for free, and we can fund development and maintenance of this from our paying customer base.