chore: disable Windows platform usage and add better warning
The Windows distribution for AWS CLI is bundled using a .msi file.
This requires special handling for unpacking, which we are not
prepared to handle/test/maintain at the moment. So instead, alert
users, point them to an issue they can +1, and in the meantime
disable CI testing for Windows so we can stay green.