feat: add `--output_json` to `s3_sync()` (#105)
This flag is not meant to be set in a `BUILD.bazel` file, but
when `bazel run`ning the target to write out a JSON file containing
which files where uploaded where and had which sha256 sum.

This makes in easy for CI jobs uploading artifacts to access that
information without having to dig into the bazel output tree.