feat: introduce `--skip_existing`: skip uploading if S3 path already exists
Fixes #84.

It is debatable if this feature should even exists given a fully
hermetic and deterministic bazel build. Unfortunately the world
is not perfect and a CI pipeline might rebuild some actions when
triggered on a previously already build commit. This would result
in overwriting S3 build artifacts which might not be preferable.